+++ R/components.R
@@ -3,20 +3,16 @@
 #' Decompose a firmly applied function (i.e., a function created by [firmly()]
 #' or [fasten()]):
 #' - `firm_core()` extracts the underlying \dQuote{core} function—the function
-#'   that is called when all arguments are valid.
-#' - `firm_checks()` extracts the checks.
+#'   that is called when all arguments are valid
 #' - `firm_error()` extracts the subclass of the error condition that is
-#'   signaled when an input validation error occurs.
+#'   signaled when an input validation error occurs
 #'
 #' @details For primitive functions, `firm_core()` returns the underlying
 #'   function as a closure.
 #' @param x Object to decompose.
 #' @return If `x` is a firmly applied function:
-#'   - `firm_core()` returns a function.
-#'   - `firm_checks()` returns a data frame with components `pred`
-#'     (quosure), `expr` (quosure), `call` (character), `msg`
-#'     (character), `is_msg_gbl` (logical), `env_msg` (environment).
-#'   - `firm_error()` returns a character vector.
+#'   - `firm_core()` returns a function
+#'   - `firm_error()` returns a character vector
 #'
 #'   In the absence of the component to be extracted, these functions return
 #'   `NULL` (in particular, whenever `x` is not of class `firm_closure`).
@@ -26,7 +22,6 @@
 #' f <- function(x, y, ...) NULL
 #' ff <- firmly(f, is.numeric, {isTRUE(. > 0)} ~ vld(x, y - x))
 #' identical(firm_core(ff), f)  # [1] TRUE
-#' firm_checks(ff)              # 4 x 6 data frame
 #' firm_error(ff)               # [1] "inputValidationError"
 #'
 #' @name components
@@ -44,8 +39,6 @@ firm_core <- firm_closure_extractor("f")
 
 #' @rdname components
 #' @export
-firm_checks <- firm_closure_extractor("chks")
-
-#' @rdname components
-#' @export
 firm_error <- firm_closure_extractor("error_class")
+
+firm_checks <- firm_closure_extractor("chks")